Drop all blank markers, section renders br
Before this patch, Content-Kit would insert blank markers in any empty
section to ensure a cursor could be placed there. This changes that to
be a section responsibility. When there are no child markers, a markup
section will create a br.

This means throughout the codebase sections and section offset should be
used to manage editing actions. There may be no marker at a given cursor
position, only using the offset in a section is reliable.

Also ensures correct behavior when deleting two sections in their
entirety.