Coalesce markers that have identical markups
This prevents some drift in the mobiledoc that comes with a lot of
editing. Prior to this, when contiguous markers were changed so that their
markups were identically, they continued to be stored as separate
markers (and rendered as contiguous textNodes). The resulted in a messy
rendered mobiledoc, with many vacuously similar markers.

Keeping markers coordinated should also reduce memory pressure in large
documents.