Return "moved" sections
To scroll to a moved card, the consumer may want access to the AT node
that has been created by the move operation.

It strikes me working on this code that "move" is a poor name for a
method that may actually destroy the node passed to it. We should
improve the API around this.