Correctly interpret a reported selection of the editor element
In Firefox, typing Command-A changes the selection to have the
editor element as the focus and anchor node (other browsers select the
first and last text nodes in the editor element). Changes
Position.fromNode to correctly detect this and return the correct
Position

Also some cleanup in tests and removed usage of `isMarkerable` functions
throughout the codebase in favor of `section.isMarkerable` property.

fixes #215