handle typing at end of atom better. wip. needs more testing
Change cursor#_findNodeForPosition to target the textnode after an
atom's ending zwnj when there is one.
Also changes the editor keyDown handler to reposition the cursor onto
that next text node.
Changes to dom parsing to read text content out of the zwnj on either
side of an atom and merge it into the before/after marker (or create new
marker(s) to accept that text)
Changes lifecycle callbacks #addCallbackOnce method to remove the
once-added callback after flushing that queue