[BUGFIX] Do not handle key events when the editor has no cursor
It is possible for the editor's element to be focused but not have a
cursor (selection). In this case, key events will fire on the editor's
element but the editor will not be able to determine a logical
`Position` (since there is no selection). When this happens, we abort
handling the key event.

fixes #344