Tweak postEditor#addMarkupToRange to consider existing markups
Fixes #360

We insert the new markup at a consistent index across the range.
If we just push on the end of the list, it can end up in different positions
of the markup stack. This results in unnecessary closing and re-opening of
the markup each time it changes position.
If we just push it at the beginning of the list, this causes unnecessary closing
and re-opening of surrounding tags.
So, we look for any tags open across the whole range, and push into the stack
at the end of those.