Add forward/backward inclusivity rules for markups fix #402 fix #392
Changes `Post#markupsInRange` to check the markers to the left and right
of the position when range is collapsed. The markups in those markers
provide `isForwardInclusive` and `isBackwardInclusive` methods that are
used to determine which markups should be considered "active" at that
range.

Also updates some test suites that weren't using Helpers.test and
Helpers.module.

fix #392
fix #402