Add Editor#beforeToggleMarkup hook
The hook runs before editor#toggleMarkup will apply its changes. If any
hook returns `false`, the changes will not be applied.

This can be used to apply arbitrary user-defined rules to reject certain
types of formatting changes, or e.g. validate a URL before allowing a
writer to create a link to it.