Add deprecation around DOM / Text Renderer removal from mobiledoc-kit
In the next major bump I believe we can remove our dependency on
Domrenderer and TextRenderer here with little downsides.  The standalone
repositories would continue to exist as they do today this function
would be only for the `mobiledoc` format which will return a "mobiledoc"
that can be passed into the standalone Renderers to reterieve their
output.

The main advantage this is that it decouples us from these renderers.

The work that'd need to happen to land this is as follows:

- [ ] remove the `format` switch and update the assertion to throw when
anything other than "mobiledoc" is passed to this function
- [ ] Update the documentation to show examples of how to do this
properly with the standalone renderers