Initial pass to get attributes in the right location
This sets up the:

- PostNodeBuilder - to pass the attributes through to the MarkupSection
constructor
- 031 Renderer - so that we can properly serialize the mobiledoc
attributes on the section
- 031 Parser - so we can take the above serialized mobiledoc
representation and parse it back into a mobiledoc post
- editor-dom - So that we can actually do something with the attribute
information

It also adds an `_attributable` function that adds our attribute stuff
here.  This will also likely be used inside list sections so that we can
also center lists made outside of markup.

The test with the `debugger` statements in it utilizes all of the above.
At this point we need to:

- Make the section element (in editor dom) actually encode
`data-md-text-align` into an appropriate inline style attribute.
- Update the DOM / Text Renderer
- Flesh out the test some more
- And verify the rest of the build passes