In the 0.3.2 renderer, omit empty section attribute arrays 🍭
Previously, the renderer was always adding a third attributes array to serialized section arrays. In the majority of cases, these are empty arrays that can be safely omitted, saving file size. Now, like with marker attributes, an array is only emitted if there is at least one custom attribute set.